show numbers on screen with click and clear screen
in Javascript file i added four different functions to do the math and a
function to check the operation asked and do the math.

I added events to number buttons to show them on screen after clicking.
I added evet to th AC button to clear the screen to 0 and reset the current value